More subpart fixes
Basically, because operators in Blender are stupid, you cannot pass them a specific object to duplicate. So in some cases, multiple objects would be instanced per object passed to it. The new code goes through all the created objects and removes the ones that don't belong again before continuing the instancing.

This addresses #99

I'm going to leave the debug output in for now, just in case further issues pop up.